requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pytest==8.1.1
httpx==0.27.0
python-dotenv==1.2.2

